+++ src/app/(auth)/login/page.tsx
@@ -3,7 +3,7 @@ import Link from 'next/link'
 
 import { cn } from '@/lib/utils'
 import { buttonVariants } from '@/components/ui/button'
-import { UserAuthForm } from '@/components/user-auth-form'
+import { LogInForm } from '@/components/LogInForm'
 
 export const metadata: Metadata = {
   title: 'Login',
@@ -31,7 +31,7 @@ export default function LoginPage() {
             Enter your email to sign in to your account
           </p>
         </div>
-        <UserAuthForm />
+        <LogInForm />
         <p className="px-8 text-center text-sm text-muted-foreground">
           <Link
             href="/register"
