+++ src/app/(auth)/register/page.tsx
@@ -3,7 +3,7 @@ import Link from 'next/link'
 import { cn } from '@/lib/utils'
 import { buttonVariants } from '@/components/ui/button'
 import { Icons } from '@/components/icons'
-import { UserAuthForm } from '@/components/user-auth-form'
+import { RegisterForm } from '@/components/RegisterForm'
 
 export const metadata = {
   title: 'Create an account',
@@ -34,7 +34,7 @@ export default function RegisterPage() {
               Enter your email below to create your account
             </p>
           </div>
-          <UserAuthForm />
+          <RegisterForm />
         </div>
       </div>
     </div>
