+++ src/components/RegisterForm.tsx
@@ -1,0 +1,164 @@
+'use client'
+
+import * as React from 'react'
+import { zodResolver } from '@hookform/resolvers/zod'
+import { useForm } from 'react-hook-form'
+import * as z from 'zod'
+
+import { cn } from '@/lib/utils'
+import { signUpSchema } from '@/lib/validations/auth'
+import { buttonVariants } from '@/components/ui/button'
+import { Input } from '@/components/ui/input'
+import { Label } from '@/components/ui/label'
+import { toast } from '@/components/ui/use-toast'
+import { Icons } from '@/components/icons'
+import { createClientComponentClient } from '@supabase/auth-helpers-nextjs'
+import { OAuthButton } from './OAuthButton'
+
+interface SignUpFormProps extends React.HTMLAttributes<HTMLDivElement> {}
+
+type FormData = z.infer<typeof signUpSchema>
+
+export function RegisterForm({ className, ...props }: SignUpFormProps) {
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+  } = useForm<FormData>({
+    resolver: zodResolver(signUpSchema),
+  })
+  const [isLoading, setIsLoading] = React.useState<boolean>(false)
+  const supabase = createClientComponentClient<Database>()
+
+  const handleSignUp = async (data: FormData) => {
+    setIsLoading(true)
+    try {
+      const { error } = await supabase.auth.signUp({
+        email: data.email,
+        password: data.password,
+        options: {
+          data: {
+            first_name: data.firstName,
+            last_name: data.lastName,
+          },
+        },
+      })
+      if (error) {
+        console.error('Sign-up error:', error)
+        return toast({
+          title: 'Error',
+          description: error.message,
+          variant: 'destructive',
+        })
+      }
+      return toast({
+        title: 'Check your email',
+        description:
+          'We sent you a login link. Be sure to check your spam too.',
+      })
+    } catch (error) {
+      console.error('Error:', error)
+    } finally {
+      setIsLoading(false)
+    }
+  }
+
+  return (
+    <div className={cn('grid gap-6', className)} {...props}>
+      <form onSubmit={handleSubmit(handleSignUp)}>
+        <div className="grid gap-2">
+          <div className="grid gap-1">
+            <Label className="sr-only" htmlFor="firstName">
+              First Name
+            </Label>
+            <Input
+              id="firstName"
+              placeholder="First Name"
+              autoCapitalize="none"
+              disabled={isLoading}
+              {...register('firstName')}
+            />
+            {errors?.firstName && (
+              <p className="px-1 text-xs text-red-600">
+                {errors.firstName.message}
+              </p>
+            )}
+          </div>
+          <div className="grid gap-1">
+            <Label className="sr-only" htmlFor="lastName">
+              Last Name
+            </Label>
+            <Input
+              id="lastName"
+              placeholder="Last Name"
+              autoCapitalize="none"
+              disabled={isLoading}
+              {...register('lastName')}
+            />
+            {errors?.lastName && (
+              <p className="px-1 text-xs text-red-600">
+                {errors.lastName.message}
+              </p>
+            )}
+          </div>
+          <div className="grid gap-1">
+            <Label className="sr-only" htmlFor="email">
+              Email
+            </Label>
+            <Input
+              id="email"
+              placeholder="name@example.com"
+              type="email"
+              autoCapitalize="none"
+              autoComplete="email"
+              autoCorrect="off"
+              disabled={isLoading}
+              {...register('email')}
+            />
+            {errors?.email && (
+              <p className="px-1 text-xs text-red-600">
+                {errors.email.message}
+              </p>
+            )}
+          </div>
+          <div className="grid gap-1">
+            <Label className="sr-only" htmlFor="password">
+              Password
+            </Label>
+            <Input
+              id="password"
+              placeholder="Password"
+              type="password"
+              autoCapitalize="none"
+              autoComplete="current-password"
+              disabled={isLoading}
+              {...register('password')}
+            />
+            {errors?.password && (
+              <p className="px-1 text-xs text-red-600">
+                {errors.password.message}
+              </p>
+            )}
+          </div>
+          <button className={cn(buttonVariants())} disabled={isLoading}>
+            {isLoading && (
+              <Icons.spinner className="mr-2 h-4 w-4 animate-spin" />
+            )}
+            Sign Up with Email
+          </button>
+          <div className="relative">
+            <div className="absolute inset-0 flex items-center">
+              <span className="w-full border-t" />
+            </div>
+            <div className="relative flex justify-center text-xs uppercase">
+              <span className="bg-background px-2 text-muted-foreground">
+                Or continue with
+              </span>
+            </div>
+          </div>
+          <OAuthButton />
+        </div>
+      </form>
+    </div>
+  )
+}
